refactor: move EditableAccordion/Section components into editable
