+++ src/explore-education-statistics-admin/src/components/editable/EditableAccordion.module.scss
@@ -1,4 +1,4 @@
-@import '~explore-education-statistics-common/src/styles/govuk-base';
+@import 'node_modules/explore-education-statistics-common/src/styles/govuk-base';
 
 .error {
   background: $govuk-error-colour;
