+++ src/explore-education-statistics-admin/src/components/editable/EditableAccordionSection.module.scss
